# Speed up finalizers for objects without object ID